gui_qt: add support for machine option plugins
A machine option plugin is a Qt widget that should follow the
interface for GUI configuration options. When configuring a machine,
the machine class (and its ancestor) will determine the plugin names
that will be searched for the widget to use.

For example, when configuring the keyboard machine, the following list
of names will be used to look for a compatible configuration widget:
- plover.machine.keyboard:Keyboard
- Keyboard
- plover.machine.base:StenotypeBase
- StenotypeBase
- builtins:object
