+++ plover/gui_qt/config_window.py
@@ -23,9 +23,6 @@ from PyQt5.QtWidgets import (
     QWidget,
 )
 
-from serial import Serial
-from serial.tools.list_ports import comports
-
 from plover.config import MINIMUM_OUTPUT_CONFIG_UNDO_LEVELS
 from plover.machine.base import SerialStenotypeBase
 from plover.machine.keyboard import Keyboard
@@ -34,8 +31,6 @@ from plover.registry import registry
 
 from plover.gui_qt.config_window_ui import Ui_ConfigWindow
 from plover.gui_qt.config_file_widget_ui import Ui_FileWidget
-from plover.gui_qt.config_keyboard_widget_ui import Ui_KeyboardWidget
-from plover.gui_qt.config_serial_widget_ui import Ui_SerialWidget
 from plover.gui_qt.utils import WindowState
 
 
@@ -123,107 +118,6 @@ class FileOption(QWidget, Ui_FileWidget):
 
     def on_path_edited(self):
         self.valueChanged.emit(expand_path(self.path.text()))
-
-
-class KeyboardOption(QWidget, Ui_KeyboardWidget):
-
-    valueChanged = pyqtSignal(QVariant)
-
-    def __init__(self):
-        super(KeyboardOption, self).__init__()
-        self.setupUi(self)
-        self.arpeggiate.setToolTip(_(
-            'Arpeggiate allows using non-NKRO keyboards.\n'
-            '\n'
-            'Each key can be pressed separately and the\n'
-            'space bar is pressed to send the stroke.'
-        ))
-        self._value = {}
-
-    def setValue(self, value):
-        self._value = value
-        self.arpeggiate.setChecked(value['arpeggiate'])
-
-    def on_arpeggiate_changed(self, value):
-        self._value['arpeggiate'] = value
-        self.valueChanged.emit(self._value)
-
-
-class SerialOption(QWidget, Ui_SerialWidget):
-
-    valueChanged = pyqtSignal(QVariant)
-
-    def __init__(self):
-        super(SerialOption, self).__init__()
-        self.setupUi(self)
-        self._value = {}
-
-    def setValue(self, value):
-        self._value = value
-        port = value['port']
-        if port is None or port == 'None':
-            self.on_scan()
-        else:
-            self.port.setCurrentText(port)
-        self.baudrate.addItems(map(str, Serial.BAUDRATES))
-        self.baudrate.setCurrentText(str(value['baudrate']))
-        self.bytesize.addItems(map(str, Serial.BYTESIZES))
-        self.bytesize.setCurrentText(str(value['bytesize']))
-        self.parity.addItems(Serial.PARITIES)
-        self.parity.setCurrentText(value['parity'])
-        self.stopbits.addItems(map(str, Serial.STOPBITS))
-        self.stopbits.setCurrentText(str(value['stopbits']))
-        timeout = value['timeout']
-        if timeout is None:
-            self.use_timeout.setChecked(False)
-            self.timeout.setValue(0.0)
-            self.timeout.setEnabled(False)
-        else:
-            self.use_timeout.setChecked(True)
-            self.timeout.setValue(timeout)
-            self.timeout.setEnabled(True)
-        self.xonxoff.setChecked(value['xonxoff'])
-        self.rtscts.setChecked(value['rtscts'])
-
-    def _update(self, field, value):
-        self._value[field] = value
-        self.valueChanged.emit(self._value)
-
-    def on_scan(self):
-        self.port.clear()
-        self.port.addItems(sorted(x[0] for x in comports()))
-
-    def on_port_changed(self, value):
-        self._update('port', value)
-
-    def on_baudrate_changed(self, value):
-        self._update('baudrate', int(value))
-
-    def on_bytesize_changed(self, value):
-        self._update('baudrate', int(value))
-
-    def on_parity_changed(self, value):
-        self._update('parity', value)
-
-    def on_stopbits_changed(self, value):
-        self._update('stopbits', float(value))
-
-    def on_timeout_changed(self, value):
-        self._update('timeout', value)
-
-    def on_use_timeout_changed(self, value):
-        if value:
-            timeout = self.timeout.value()
-        else:
-            timeout = None
-        self.timeout.setEnabled(value)
-        self._update('timeout', timeout)
-
-    def on_xonxoff_changed(self, value):
-        self._update('xonxoff', value)
-
-    def on_rtscts_changed(self, value):
-        self._update('rtscts', value)
 
 
 class KeymapOption(QTableWidget):
@@ -506,15 +400,22 @@ class ConfigWindow(QDialog, Ui_ConfigWindow, WindowState):
         self.finished.connect(self.save_state)
 
     def _machine_option(self, *args):
+        machine_options = {
+            plugin.name: plugin.obj
+            for plugin in registry.list_plugins('gui.qt.machine_option')
+        }
         machine_type = self._config['machine_type']
         machine_class = registry.get_plugin('machine', machine_type).obj
-        if issubclass(machine_class, Keyboard):
-            opt_class = KeyboardOption
-        elif issubclass(machine_class, SerialStenotypeBase):
-            opt_class = SerialOption
-        else:
-            opt_class = NopeOption
-        return opt_class(*args)
+        for klass in machine_class.mro():
+            # Look for `module_name:class_name` before `class_name`.
+            for name in (
+                '%s:%s' % (klass.__module__, klass.__name__),
+                klass.__name__,
+            ):
+                opt_class = machine_options.get(name)
+                if opt_class is not None:
+                    return opt_class(*args)
+        return NopeOption(*args)
 
     def _update_keymap(self, machine_type=None, system_name=None):
         if machine_type is None:
