+++ setup.cfg
@@ -61,6 +61,9 @@ plover.dictionary =
 plover.gui =
 	none = plover.gui_none.main
 	qt   = plover.gui_qt.main
+plover.gui.qt.machine_option =
+	plover.machine.base:SerialStenotypeBase = plover.gui_qt.machine_options:SerialOption
+	plover.machine.keyboard:Keyboard        = plover.gui_qt.machine_options:KeyboardOption
 plover.gui.qt.tool =
 	add_translation = plover.gui_qt.add_translation_dialog:AddTranslationDialog
 	lookup          = plover.gui_qt.lookup_dialog:LookupDialog
